requirements.txt
python-slugify==3.0.2
hdx-python-api==3.8.2
-r docker-requirements.txt
